Fix cling compilation error in ParaviewAdaptorTest.GenerateSimulationInfoJson after merge
Visualization adaptor library was not loaded in the constructor of
Scheduler anymore.
